handle empty data and create component empty data
