+++ features/home/index.tsx
@@ -2,9 +2,9 @@ import { Container } from "@mui/system";
 import Carrousel from "@/components/Carrousel";
 import { useQuery } from "react-query";
 import { getSlideshow } from "services/slideshow";
-import Featured from "./featured";
-import NewListings from "./new-listing";
-import Events from "./event";
+import Featured from "./Featured";
+import NewListings from "./NewListing";
+import Events from "./Event";
 
 const MainPage = () => {
   const { data: slideshowData, isSuccess } = useQuery(["getSlideshow"], () =>
