+++ pages/[game]/index.tsx
@@ -1,4 +1,4 @@
-import { Items, Reviews, TopPlayers } from "@/features/game/overview";
+import { Items, Reviews, TopPlayers } from "@/features/game/Overview";
 
 export default function OverviewPages() {
   return (
