+++ pages/[game]/mint.tsx
@@ -1,4 +1,4 @@
-import Mints from "@/features/game/Mint/Mints";
+import Mints from "@/features/game/Mint";
 
 export default function MintPage() {
   return <Mints />
