+++ styles/global.scss
@@ -88,3 +88,10 @@ svg {
     width: 283px !important
   }
 }
+
+.featured-swipperSmDown  {
+  .swiper-slide {
+    transform: none;
+    width: 283px !important
+  }
+}
